+++ sql/promscale--0.0.0.sql
@@ -407,10 +407,10 @@ BEGIN
         ORDER BY m.table_schema, m.table_name
     )
     LOOP
-        EXECUTE format($sql$ALTER TABLE %I.%I OWNER TO prom_admin $sql$, _rec.table_schema, _rec.table_name);
-        EXECUTE format($sql$ALTER TABLE prom_data_series.%I OWNER TO prom_admin $sql$, _rec.series_table);
-        EXECUTE format($sql$ALTER VIEW prom_series.%I OWNER TO prom_admin $sql$, _rec.series_table);
-        EXECUTE format($sql$ALTER VIEW prom_metric.%I OWNER TO prom_admin $sql$, _rec.table_name);
+        EXECUTE format($sql$ALTER TABLE IF EXISTS %I.%I OWNER TO prom_admin $sql$, _rec.table_schema, _rec.table_name);
+        EXECUTE format($sql$ALTER TABLE IF EXISTS prom_data_series.%I OWNER TO prom_admin $sql$, _rec.series_table);
+        EXECUTE format($sql$ALTER VIEW IF EXISTS prom_series.%I OWNER TO prom_admin $sql$, _rec.series_table);
+        EXECUTE format($sql$ALTER VIEW IF EXISTS prom_metric.%I OWNER TO prom_admin $sql$, _rec.table_name);
    END LOOP;
 END
 $block$;
@@ -427,7 +427,7 @@ $block$;
          ORDER BY e.table_name
      )
      LOOP
-         EXECUTE format($sql$ALTER TABLE prom_data_exemplar.%I OWNER TO prom_admin $sql$, _rec.table_name);
+         EXECUTE format($sql$ALTER TABLE IF EXISTS prom_data_exemplar.%I OWNER TO prom_admin $sql$, _rec.table_name);
      END LOOP;
  END;
  $block$